pyproject.toml
[project]
name = "circuitbreaker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ipynotebook>=0.1.0",
    "jupyter>=1.1.1",
    "pandas>=2.3.1",
    "python-dateutil>=2.9.0.post0",
    "requests>=2.32.4",
    "tenacity>=9.1.2",
]
